react: Split FilterToggles into separate component
